Changed interface

+++ price_calculator/price_calculator.py
@@ -1,3 +1,3 @@
 class PriceCalculator:
-    def order_value(self, count: int, price: int) -> int:
+    def order_value(self, count: int, price: float) -> float:
         return count * price
